+++ public/locales/en/translations.json
@@ -2,6 +2,8 @@
     "fr-CA": "Français",
     "en-CA": "English",
 
+    "index-desc" : "Jean-Michel Viel, musician, educator and programmer.",
+
     "about-url": "about",
     "music-url": "music",
     "computer-science-url": "computer-science",
