+++ public/locales/fr/translations.json
@@ -2,6 +2,8 @@
     "fr-CA": "Français",
     "en-CA": "English",
 
+    "index-desc" : "Jean-Michel Viel, musicien, pédagogue et programmeur.",
+
     "about-url": "a-propos",
     "music-url": "musique",
     "computer-science-url": "informatique",
