+++ wrangler.toml
@@ -1,8 +1,10 @@
 workers_dev = true
-name = "my-react-router-worker"
+name = "jmviel-siteweb"
 compatibility_date = "2024-11-18"
 main = "./build/server/index.js"
 assets = { directory = "./build/client/" }
+
+compatibility_flags = [ "nodejs_compat" ]
 
 [vars]
 VALUE_FROM_CLOUDFLARE = "Hello from Cloudflare"
